added syncing to the project with some test lines; next I need to write some tests and hook it up to the actual sync commands
