+++ openmrs-client/src/main/java/org/openmrs/mobile/activities/login/LoginActivity.java
@@ -14,13 +14,30 @@
 
 package org.openmrs.mobile.activities.login;
 
+import android.accounts.Account;
+import android.accounts.AccountManager;
+import android.content.ContentResolver;
+import android.content.Context;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.Menu;
 
 import org.openmrs.mobile.R;
 import org.openmrs.mobile.activities.ACBaseActivity;
 
 public class LoginActivity extends ACBaseActivity {
+
+	// Constants
+	// The authority for the sync adapter's content provider
+	public static final String AUTHORITY = "org.openmrs.mobile.provider";
+	// An account type, in the form of a domain name
+	public static final String ACCOUNT_TYPE = "org.openmrs.mobile.datasync";
+	// The account name (this isn't used by the app, but is needed for syncing)
+	public static final String ACCOUNT = "defaultAccount";
+	// Sync interval constants
+	public static final long SECONDS_PER_MINUTE = 60L;
+	public static final long SYNC_INTERVAL_IN_MINUTES = 1L;
+	public static final long SYNC_INTERVAL = SYNC_INTERVAL_IN_MINUTES * SECONDS_PER_MINUTE;
 
 	public LoginContract.Presenter mPresenter;
 
@@ -42,11 +59,51 @@ public class LoginActivity extends ACBaseActivity {
 
 		mPresenter = new LoginPresenter(loginFragment, mOpenMRS);
 
+		turnOnSyncing();
+
 	}
 
 	@Override
 	public boolean onCreateOptionsMenu(Menu menu) {
 		super.onCreateOptionsMenu(menu);
 		return true;
+	}
+
+	private void turnOnSyncing() {
+		// Create the dummy account
+		Account mAccount = createSyncAccount(this);
+		// Turn on periodic syncing
+		ContentResolver.setSyncAutomatically(mAccount, AUTHORITY, true);
+		ContentResolver.addPeriodicSync(mAccount, AUTHORITY, Bundle.EMPTY, 3600);
+	}
+
+	/**
+	 * Create a new dummy account for the sync adapter
+	 *
+	 * @param context The application context
+	 */
+	public static Account createSyncAccount(Context context) {
+		// Create the account type and default account
+		Account newAccount = new Account(ACCOUNT, ACCOUNT_TYPE);
+		// Get an instance of the Android account manager
+		AccountManager accountManager = (AccountManager) context.getSystemService(ACCOUNT_SERVICE);
+
+		// Below needed for syncing, but is for dummy account
+		if (accountManager.addAccountExplicitly(newAccount, null, null)) {
+            /*
+             * If you don't set android:syncable="true" in
+             * in your <provider> element in the manifest,
+             * then call context.setIsSyncable(account, AUTHORITY, 1)
+             * here.
+             */
+			// Intentionally left blank
+		} else {
+            /*
+             * The account exists or some other error occurred. Log this, report it,
+             * or handle it internally.
+             */
+			// Intentionally left blank
+		}
+		return newAccount;
 	}
 }
