+++ openmrs-client/src/main/res/xml/syncadapter.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <sync-adapter
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:contentAuthority="org.openmrs.mobile.data.sync"
-    android:accountType="com.android.example.datasync"
+    android:contentAuthority="org.openmrs.mobile.provider"
+    android:accountType="org.openmrs.mobile.datasync"
     android:userVisible="false"
-    android:supportsUploading="false"
+    android:supportsUploading="true"
     android:allowParallelSyncs="false"
     android:isAlwaysSyncable="true" />
